requirements.txt
dwhutils==0.1.9
python-dotenv
ConfigParser
psycopg2-binary
pandas
sqlalchemy
requests
pymongo
pandasql
termcolor2
pymysql
apache-airflow
connection
stem
PyYaml
marquez-python
marquez-airflow
sshtunnel
swifter
pykeepass
matplotlib
statsmodels
seaborn
scikit-learn
pika
pmdarima
dwhutils